Create RGB and Timer
I have this code and I'm using an Arduino Mega. The code is to have an IR Remote use the power button and the play/pause button. It should be defined well enough as to what it is supposed to do. When I press play, power goes to the rgb led and it turns on and then the led display turns on and shows all zeros. When I press play, the rgb changes color and the timer begins to count up to 60 and back down to 0. However, when I press play, the display shows some random things really fast and then shows one zero and doesn't display the timer counting up or down. What can I do??